size_t

+++ search_algorithms/0-linear.c
@@ -10,7 +10,7 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	int x;
+	size_t x;
 
 	if (array == NULL)
 		return (-1);
